backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.10
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
alembic==1.13.1
python-dotenv==1.0.1
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
qrcode==7.4.2
pillow==10.2.0
pytest==8.0.1
pytest-cov==4.1.0
httpx==0.26.0
python-dateutil==2.8.2